Refactor Maps Activity into a MapsFragment
