+++ app/src/main/AndroidManifest.xml
@@ -23,14 +23,20 @@
         android:theme="@style/Theme.Qrchive"
         tools:targetApi="31">
 
+        <!--
+             TODO: Before you run your application, you need a Google Maps API key.
+
+             To get one, follow the directions here:
+
+                https://developers.google.com/maps/documentation/android-sdk/get-api-key
+
+             Once you have your API key (it starts with "AIza"), define a new property in your
+             project's local.properties file (e.g. MAPS_API_KEY=Aiza...), and replace the
+             "YOUR_API_KEY" string in this file with "${MAPS_API_KEY}".
+        -->
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="@string/GOOGLE_MAPS_API_KEY" />
-
-        <activity
-            android:name=".Activities.MapsActivity"
-            android:exported="false"
-            android:label="@string/title_activity_maps" />
+            android:value="AIzaSyBjBKxQPoyQ5OEA8NL9x08LflycdZrwrmk" />
         <activity
             android:name=".Activities.MainActivity"
             android:exported="true">
