--- app/src/main/res/values/dimens.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <dimen name="text_margin">16dp</dimen>
-</resources>
